requirements.txt
accelerate==1.6.0
anyascii==0.3.2
certifi==2025.1.31
charset-normalizer==3.4.1
defusedxml==0.7.1
diffusers==0.33.1
filelock==3.18.0
fsspec==2025.3.0
h5py==3.13.0
huggingface-hub==0.29.3
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
Jinja2==3.1.6
langdetect==1.0.9
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pillow==11.1.0
psutil==7.0.0
pyclipper==1.3.0.post6
pypdfium2==4.30.1
python-doctr==0.11.0
PyYAML==6.0.2
RapidFuzz==3.12.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
setuptools==75.8.2
shapely==2.0.7
six==1.17.0
sympy==1.13.1
tifffile==2025.3.13
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.3.0
wheel==0.45.1
zipp==3.21.0
requests-oauthlib==2.0.0